Fix line alignment issues and add option to repeat line
Everything in this works. The changes are explained below

- lines 52-55 The `.0` is not necessary to prevent truncation of the
result
- `repeatLineTimes` is the number of times to draw each line of pixels.
This is useful for using a pen that is not thick enough compared to the
number of horizontal lines in the image.
- lines 182, 187, 200 the conditional operator for the x position
offsets everything by `-step` when step is negative this maintains
vertical alignment of the pixels when drawn while moving in both
directions.
- line 197 This if statement moves the pen over to finish drawing the
last pixel in the line. This also raises the pen. This will result in
the end of each line not being connected.